nova versao

+++ src/main/java/CRUD.java
@@ -17,7 +17,15 @@ public class CRUD {
       em.getTransaction().begin();
       em.persist(m1);
       em.getTransaction().commit();
+      genero.addMidia(m1);
+      diretor.addMidia(m1);
+      em.getTransaction().begin();
+      em.merge(genero);
+      em.merge(diretor);
+      em.getTransaction().commit();
       em.close();
+
+
   }
 
   public Midia searchMidia(int ID){
@@ -34,6 +42,7 @@ public class CRUD {
       Midia m1 = searchMidia(ID);
       System.out.println("Digite 1 para atualizar o nome ou 2 para o ano de lançamento");
       int resposta = scan.nextInt();
+      scan.nextLine();
       if(resposta==1){
           System.out.println("Digite o novo nome: ");
           String novoNome = scan.nextLine();
@@ -52,17 +61,49 @@ public class CRUD {
   }
   public void deleteMidia(int ID){
       Scanner scan = new Scanner(System.in);
-      Midia m1 = searchMidia(ID);
-      System.out.println("Confirmar exclusão? [y/n]");
-      String resposta = scan.nextLine();
-      if(resposta.equals("y")){
-          EntityManager em = emf.createEntityManager();
-          em.getTransaction().begin();
-          em.remove(m1);
-          System.out.println("Mídia excluída com sucesso");
+      EntityManager em = emf.createEntityManager();
+
+      try {
+          Midia m1 = em.find(Midia.class, ID);
+
+          if (m1 != null) {
+              System.out.println("Confirmar exclusão? [y/n]");
+              String resposta = scan.nextLine();
+
+              if (resposta.equals("y")) {
+                  Genero g1 =em.find(Genero.class,m1.getIDGenero());
+                  g1.getMidias().remove(m1);
+                  em.getTransaction().begin();
+                  em.merge(g1);
+                  em.getTransaction().commit();
+
+                  Diretor d1 = em.find(Diretor.class, m1.getIDDiretor());
+                  d1.getMidias().remove(m1);
+
+                  em.getTransaction().begin();
+                  em.merge(d1);
+                  em.getTransaction().commit();
+
+                  em.getTransaction().begin();
+                  em.remove(m1);
+                  em.getTransaction().commit();
+
+                  System.out.println("Mídia excluída com sucesso");
+              } else if (resposta.equals("n")) {
+                  System.out.println("Operação cancelada");
+              } else {
+                  System.out.println("Resposta inválida");
+              }
+          } else {
+              System.out.println("Mídia não encontrada");
+          }
+      } catch (Exception e) {
+          System.out.println("Ocorreu um erro durante a exclusão: " + e.getMessage());
+          if (em.getTransaction().isActive()) {
+              em.getTransaction().rollback();
+          }
+      } finally {
           em.close();
-      } else if (resposta.equals("n")) {
-          System.out.println("Operação cancelada");
 
       }
 
@@ -87,34 +128,51 @@ public class CRUD {
     public void updateGenero(int ID){
         Scanner scan = new Scanner(System.in);
         Genero genero = searchGenero(ID);
-        System.out.println("Digite 1 para atualizar o nome do gênero");
-        int resposta = scan.nextInt();
-        if(resposta == 1){
-            System.out.println("Digite o novo nome do gênero: ");
-            String novoNome = scan.nextLine();
-            genero.setNome(novoNome);
-        }
+
+        System.out.println("Digite o novo nome do gênero: ");
+        String novoNome = scan.nextLine();
+        genero.setNome(novoNome);
+
         EntityManager em = emf.createEntityManager();
         em.getTransaction().begin();
         em.merge(genero);
         em.getTransaction().commit();
         em.close();
+        System.out.println("atualizado.");
     }
 
     public void deleteGenero(int ID){
         Scanner scan = new Scanner(System.in);
-        Genero genero = searchGenero(ID);
-        System.out.println("Confirmar exclusão do gênero? [y/n]");
-        String resposta = scan.nextLine();
-        if(resposta.equals("y")){
-            EntityManager em = emf.createEntityManager();
-            em.getTransaction().begin();
-            em.remove(genero);
-            System.out.println("Gênero excluído com sucesso");
-            em.getTransaction().commit();
+        EntityManager em = emf.createEntityManager();
+
+        try {
+            Genero g1 = em.find(Genero.class, ID);
+
+            if (g1 != null) {
+                System.out.println("Confirmar exclusão? [y/n]");
+                String resposta = scan.nextLine();
+
+                if (resposta.equals("y")) {
+                    em.getTransaction().begin();
+                    em.remove(g1);
+                    em.getTransaction().commit();
+                    System.out.println("Gênero excluída com sucesso");
+                } else if (resposta.equals("n")) {
+                    System.out.println("Operação cancelada");
+                } else {
+                    System.out.println("Resposta inválida");
+                }
+            } else {
+                System.out.println("Gênero não encontrado");
+            }
+        } catch (Exception e) {
+            System.out.println("Ocorreu um erro durante a exclusão: " + e.getMessage());
+            if (em.getTransaction().isActive()) {
+                em.getTransaction().rollback();
+            }
+        } finally {
             em.close();
-        } else if (resposta.equals("n")) {
-            System.out.println("Operação cancelada");
+
         }
     }
 
@@ -136,34 +194,50 @@ public class CRUD {
     public void updateDiretor(int ID){
         Scanner scan = new Scanner(System.in);
         Diretor diretor = searchDiretor(ID);
-        System.out.println("Digite 1 para atualizar o nome do diretor");
-        int resposta = scan.nextInt();
-        if(resposta == 1){
-            System.out.println("Digite o novo nome do diretor: ");
-            String novoNome = scan.nextLine();
-            diretor.setNome(novoNome);
-        }
+
+        System.out.println("Digite o novo nome do diretor: ");
+        String novoNome = scan.nextLine();
+        diretor.setNome(novoNome);
+
         EntityManager em = emf.createEntityManager();
         em.getTransaction().begin();
         em.merge(diretor);
         em.getTransaction().commit();
         em.close();
+        System.out.println("atualizado.");
     }
-    public void deleteDiretor(int ID){
+    public void deleteDiretor(int ID) {
         Scanner scan = new Scanner(System.in);
-        Diretor diretor = searchDiretor(ID);
-        System.out.println("Confirmar exclusão do diretor? [y/n]");
-        String resposta = scan.nextLine();
-        if(resposta.equals("y")){
-            EntityManager em = emf.createEntityManager();
-            em.getTransaction().begin();
-            em.remove(diretor);
-            System.out.println("Diretor excluído com sucesso");
-            em.getTransaction().commit();
+        EntityManager em = emf.createEntityManager();
+
+        try {
+            Diretor d1 = em.find(Diretor.class, ID);
+
+            if (d1 != null) {
+                System.out.println("Confirmar exclusão? [y/n]");
+                String resposta = scan.nextLine();
+
+                if (resposta.equals("y")) {
+                    em.getTransaction().begin();
+                    em.remove(d1);
+                    em.getTransaction().commit();
+                    System.out.println("Diretor excluída com sucesso");
+                } else if (resposta.equals("n")) {
+                    System.out.println("Operação cancelada");
+                } else {
+                    System.out.println("Resposta inválida");
+                }
+            } else {
+                System.out.println("Diretor não encontrada");
+            }
+        } catch (Exception e) {
+            System.out.println("Ocorreu um erro durante a exclusão: " + e.getMessage());
+            if (em.getTransaction().isActive()) {
+                em.getTransaction().rollback();
+            }
+        } finally {
             em.close();
-        } else if (resposta.equals("n")) {
-            System.out.println("Operação cancelada");
+
         }
     }
-
 }
